changed wsusgroups parameter to a single string
this is a breaking change.
there was an issue with the new prefetcher that complicates this
provider approach.  the issue stems from the namevar needing to
be unique, and a namevar that can be appropriately crafted from
the 'is' state to then match the state defined in ones manifest.

so now the expected structure for namevar kb123456-group
and the wsusgroup => 'group'

This will ensure you don't duplicate namevar.